int8 -> int16 for maps

+++ terra/map_utils/jax_terrain_generation.py
@@ -2,7 +2,7 @@ import jax
 import jax.numpy as jnp
 from jax.scipy.signal import convolve2d
 
-IntMap = jnp.int8  # TODO import
+IntMap = jnp.int16  # TODO import
 
 
 def generate_clustered_bitmap(
